Condicional de que si eres producer, no salga el boton de convertirte en producer
